fix: rename omegaconf
